registration/requirement.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
bson==0.5.8
celery==4.4.2
dataclasses==0.6
Django==2.2.9
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djongo==1.2.36
kombu==4.6.8
mongoengine==0.20.0
paho-mqtt==1.5.0
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.2.4
vine==1.3.0


# python 3.8
# mongo db version 4.2.6
